refactor(header-profile): moved login/logout logic into header icon
