+++ src/constants/portuguese.ts
@@ -15,17 +15,6 @@ export const portugueseTextMap = {
           `Erro na API: ${message}`
         ),
       },
-      profile: {
-        loading: 'Carregando informações do perfil...',
-        noUser: 'Nenhum usuário conectado',
-        title: 'Informações do Perfil',
-        noEmail: 'Email não encontrado',
-        email: ({ email }: { email: string }) => `Email: ${email}`,
-        noName: 'Nome não encontrado',
-        name: ({ name }: { name: string }) => `Nome: ${name}`,
-        signInGoogle: 'Conectar com Google',
-        signOut: 'Sair'
-      }
     }
   }
 } as const;
